% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autophagy.R
\name{autophagy}
\alias{autophagy}
\alias{autophagy.SOmap}
\title{Map will eat itself (WIP)}
\usage{
autophagy(x, ...)

\method{autophagy}{SOmap}(x, ...)
}
\arguments{
\item{x}{object to plot.}

\item{...}{other plotting options.}
}
\description{
TODO:
 - deconstruct SOauto_map to separate the plotting and aesthetics handling from the data prep
 - etc.
}
\examples{
\dontrun{x <- SOauto_map()
autophagy(a)}
}
