% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap_data.R
\docType{data}
\name{bathy_grob}
\alias{bathy_grob}
\title{Bathymetric data for maps}
\format{An object of class \code{"rastergrob"}}
\source{
\href{https://www.gebco.net/data_and_products/gridded_bathymetry_data/}{GEBCO}
}
\usage{
data(bathy_grob)
}
\description{
As for data Bathy, but in the form of a rastergrob. Bathymetric data reprocessed from the GEBCO_2014 Grid data set.
}
\references{
The GEBCO_2014 Grid, version 20150318
}
\keyword{datasets}
