% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.discrete.r
\name{is.discrete}
\alias{is.discrete}
\title{used in SOleg to test color palettes}
\usage{
is.discrete(x)
}
\arguments{
\item{x}{Object to test for is it discrete}
}
\value{
returns true false
}
\description{
used in SOleg to test color palettes
}
