% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOmap.R
\name{NOmap}
\alias{NOmap}
\title{Northern Ocean Map}
\usage{
NOmap(Bathleg = TRUE, Border = TRUE, Trim = 45, Grats = FALSE,
  straight = FALSE, land = TRUE, fronts = FALSE,
  frontcols = c("hotpink", "orchid", "plum"), bordercol = c("white",
  "black"), borderwidth = 2, gratcol = "grey70")
}
\arguments{
\item{Bathleg}{logical: if \code{TRUE}, insert the bathymetry legend}

\item{Border}{logical: if \code{TRUE}, insert longitude border}

\item{Trim}{numeric: latitude to trim map to, set to 10 for effectively no trim}

\item{Grats}{logical: if \code{TRUE}, insert graticule grid}

\item{straight}{logical: if \code{TRUE}, leave a blank space on the side for a straight legend}

\item{land}{logical: if \code{TRUE}, plot coastline}

\item{fronts}{logical: if \code{TRUE}, plot ocean fronts (Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front)}

\item{frontcols}{character: colours for fronts}

\item{bordercol}{character: colours for longitude border}

\item{borderwidth}{numeric: thickness (in degrees of latitude) of the border}

\item{gratcol}{string: colour for graticule grid}
}
\value{
Produces at the very least a round bathymetry map of the southern hemisphere.
}
\description{
Function for creating round Northern Ocean maps.
}
\examples{
NOmap(Trim=-45, Grats=TRUE)

}
