% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOproj.R
\name{projection}
\alias{projection}
\alias{projection.SOmap}
\title{Projection metadata}
\usage{
projection.SOmap(x, asText = TRUE)
}
\arguments{
\item{x}{filename (character), Extent, Raster*, sf, SpatialPixels*, SpatialGrid*, object, 'image', matrix, im, or missing. Supported file types are the 'native' raster package format and those that can be read via \code{rgdal} (see \code{\link[rgdal]{readGDAL}}}
}
\description{
Obtain the projection used in a SOmap object.
}
