% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOproj.R
\name{reproj.SOmap}
\alias{reproj.SOmap}
\alias{reproj}
\alias{reproj.SOauto_map}
\title{Reproject SOmap}
\usage{
\method{reproj}{SOmap}(x, target, ..., source = NULL)

\method{reproj}{SOauto_map}(x, target, ..., source = NULL)
}
\arguments{
\item{x}{coordinates}

\item{target}{target specification (PROJ.4 string or epsg code)}

\item{...}{arguments passed to \code{\link[proj4:ptransform]{proj4::ptransform()}}}

\item{source}{source specification (PROJ.4 string or epsg code)}
}
\description{
Reproject a SOmap object by specifying a 'target' projection string (PROJ4)
}
\details{
See [reproj::reproj()] for details.
}
\section{Warning}{

So many ...
}

\examples{
set.seed(27)
amap <- SOauto_map()
reproj(amap, "+proj=moll")
reproj(amap, "+proj=laea +lat_0=-55 +lon_0=154 +datum=WGS84")

bmap <- SOmap(Trim = -35)

## works great!
reproj(bmap, "+proj=stere +lat_0=-90 +lon_0=147 +lat_ts=-71 +datum=WGS84")
## these aren't exactly ideal
reproj(bmap, "+proj=ortho +lat_0=-70")
reproj(bmap, "+proj=laea +lat_0=-55 +lon_0=154 +datum=WGS84")
}
\seealso{
[reproj::reproj()]
}
