% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOcrs.R
\name{SOcrs}
\alias{SOcrs}
\title{SOmap coordinate system}
\usage{
SOcrs(crs = NULL)
}
\arguments{
\item{crs}{provide PROJ string to set the value}
}
\description{
Set or return the coordinate system currently in use.
}
\details{
If argument `crs` is NULL, the function returns the current value (which may be `NULL``).
}
\examples{
\dontrun{
SOmap()
SOcrs()
}
}
