% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{SOgg}
\alias{SOgg}
\title{Generate a ggplot2 representation of an SOmap object}
\usage{
SOgg(x)
}
\arguments{
\item{x}{SOmap: object as returned by \code{SOmap} or \code{SOmap2}}
}
\value{
A ggplot object
}
\description{
Note: this function is still experimental! Use at your own risk.
}
\examples{
\dontrun{
  p <- SOmap2(Trim = -45, IWC = TRUE, IWClab = TRUE, Grats = TRUE, fronts = TRUE,
              MPA = TRUE, MPAlab = TRUE)
  SOgg(p)
}

}
