% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOleg.R
\name{SOleg}
\alias{SOleg}
\title{SOleg creating rounded legends for SOmap}
\usage{
SOleg(x = NULL, position = "topright", col = NULL, ticks = NULL,
  tlabs = NULL, breaks = NULL, Trim = -45, type = "discrete",
  label = "", ladj = 0.5, lsrt = 0, lcex = 0.75, tadj = 0.5,
  tcex = 1, rnd = NULL)
}
\arguments{
\item{x}{Object to obtain min and max values from for type='continuous' default=NULL}

\item{position}{Where you want the legend ("topleft","topright", "bottomright")}

\item{col}{Color pattern to use.}

\item{ticks}{Number of ticks to include on the legend. Only used with type='continuous'}

\item{tlabs}{Tick labels. Needed for type='discrete' optonal for type='continuous' if x is given}

\item{breaks}{Numeric vector to create legend ticks for type='continuous' if x is given eg breaks<-c(1,2,3)}

\item{Trim}{Trim that was used to create the SOmap.}

\item{type}{Type of legend to be plotted 'discrete' or 'continuous' default='discrete'}

\item{label}{Legend label.}

\item{ladj}{Distance to adjust the tick labels from the ticks. default = 0.5}

\item{lsrt}{Angle of the tick labels. default = 0}

\item{lcex}{Size of the tick labels. default = 0.75}

\item{tadj}{Distance to adjust the title from the ticks. default = 0.5}

\item{tcex}{Size of the title text. default=1}

\item{rnd}{optional rounding factor for continuous legends using the round() function. default = NULL.}
}
\value{
Creates rounded legends
}
\description{
SOleg creating rounded legends for SOmap
}
\examples{
\dontrun{
SOmap()
#Discrete Legend

SOleg(position="topleft",
      col=viridis::viridis(5),
      tlabs = c("a","b","c","d", "e"),
      Trim=-45,
      label="Species")

#Continuous Legend

nums<-runif(100)

brks<-c(0.1,0.2,0.5,0.9)

SOleg(x=nums,
      position="topright",
      col=viridis::viridis(80),
      breaks=brks
      Trim=-45,
      label="Species"
      rnd=1,
      type="continuous")
}
}
