% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap-old.R, R/SOmap.R
\name{SOmap}
\alias{SOmap}
\title{Southern Ocean Map}
\usage{
SOmap(Bathleg = TRUE, Border = TRUE, Trim = -45, Grats = FALSE,
  straight = FALSE, land = TRUE, fronts = FALSE,
  frontcols = c("hotpink", "orchid", "plum"), bordercol = c("white",
  "black"), gratcol = "grey70")

SOmap(Bathleg = TRUE, Border = TRUE, Trim = -45, Grats = FALSE,
  straight = FALSE, land = TRUE, fronts = FALSE,
  frontcols = c("hotpink", "orchid", "plum"), bordercol = c("white",
  "black"), gratcol = "grey70")
}
\arguments{
\item{Bathleg}{Insert the bathymetry legend.}

\item{Border}{Insert longitude border.}

\item{Trim}{Longitude to trim map to.}

\item{Grats}{Insert graticule grid.}

\item{straight}{Do you need a blank space on the side for a straight legend.}

\item{land}{Plot land boundary}

\item{fronts}{Plot ocean fronts: Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front}

\item{frontcols}{colors for fronts

 @return
Produces at the very base a round bathymetry map of the southern hemisphere.}

\item{bordercol}{Colors for longitude border; Default is c("black","white").}

\item{gratcol}{Color for graticule grid; Default is grey.}

\item{Bathleg}{logical: if \code{TRUE}, insert the bathymetry legend}

\item{Border}{logical: if \code{TRUE}, insert longitude border}

\item{Trim}{numeric: longitude to trim map to}

\item{Grats}{logical: if \code{TRUE}, insert graticule grid}

\item{bordercol}{character: colours for longitude border}

\item{gratcol}{string: colour for graticule grid}

\item{straight}{logical: if \code{TRUE}, leave a blank space on the side for a straight legend}

\item{land}{logical: if \code{TRUE}, plot coastline}

\item{fronts}{logical: if \code{TRUE}, plot ocean fronts (Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front)}

\item{frontcols}{character: colours for fronts}
}
\value{
Produces at the very least a round bathymetry map of the southern hemisphere.
}
\description{
Function for creating round Southern Ocean maps.

Function for creating round Southern Ocean maps.
}
\examples{
\dontrun{
Dat.Dir<-getwd()
png(paste(Dat.Dir,'/SOmap.png', sep=''), width=22, height=20, units='cm', res=600)
SOmap(Trim=-45, Grats=TRUE)
dev.off()
}
\dontrun{
  SOmap(Trim = -45, Grats = TRUE)
}
}
