% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOplot.R
\name{SOplot}
\alias{SOplot}
\title{Southern Ocean plot}
\usage{
SOplot(lon, lat, prj, crs, add = TRUE, ...)
}
\arguments{
\item{lon}{longitude object}

\item{lat}{lattitude object}

\item{prj}{starting projection (default = longlat)}

\item{crs}{target projection (default = stereo)}

\item{add}{add layer to plot (default = TRUE)}

\item{...}{other plot options}
}
\value{
Produces at the very base a round bathymetry map of the southern hemisphere.
}
\description{
Reproject and add point layers to either SOmap or SOauto_map layers.
}
\examples{
\dontrun{
 x<-c(-70, -60,-50, -90)
 y<-c(-50, -75, -45, -60)
 map<-SOauto_map(x,y, input_lines = FALSE)
 map
 SOplot(lon=y, lat=x,crs = map$projection,pch=19,col=6)
}
}
