% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOproj.R
\name{SOproj}
\alias{SOproj}
\title{Southern projection}
\usage{
SOproj(lon, lat, prj, crs, data)
}
\arguments{
\item{lon}{longitude object.}

\item{lat}{lattitude object.}

\item{prj}{starting projection (default = longlat)}

\item{crs}{target projection (default = stereo)}

\item{data}{optional data to be included}
}
\value{
Produces at the very base a round bathymetry map of the southern hemisphere.
}
\description{
Function for reprojecting data.
}
\examples{
\dontrun{
 x<-c(-70, -60,-50, -90)
 y<-c(-50, -75, -45, -60)
 pnts<-SOproj(lon = y, lat = x)
 SOmap2(CCAMLR=T)
 plot(pnts, pch=19, col=3, add=TRUE)
}
}
