% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOmap.R
\name{SOmap}
\alias{SOmap}
\title{Southern Ocean Map}
\usage{
SOmap(Bathleg = TRUE, Border = TRUE, Trim = -45, Grats = FALSE,
  straight = FALSE, land = TRUE, fronts = FALSE,
  frontcols = c("hotpink", "orchid", "plum"), bordercol = c("white",
  "black"), borderwidth = 2, gratcol = "grey70")
}
\arguments{
\item{Bathleg}{logical: if \code{TRUE}, insert the bathymetry legend}

\item{Border}{logical: if \code{TRUE}, insert longitude border}

\item{Trim}{numeric: longitude to trim map to}

\item{Grats}{logical: if \code{TRUE}, insert graticule grid}

\item{straight}{logical: if \code{TRUE}, leave a blank space on the side for a straight legend}

\item{land}{logical: if \code{TRUE}, plot coastline}

\item{fronts}{logical: if \code{TRUE}, plot ocean fronts (Subantarctic Front, Polar Front, Southern Antarctic Circumpolar Current Front)}

\item{frontcols}{character: colours for fronts}

\item{bordercol}{character: colours for longitude border}

\item{borderwidth}{numeric: thickness (in degrees of latitude) of the border}

\item{gratcol}{string: colour for graticule grid}
}
\value{
Produces at the very least a round bathymetry map of the southern hemisphere.
}
\description{
Function for creating round Southern Ocean maps.
}
\examples{
\dontrun{
tfile <- tempfile("SOmap", fileext = ".png")
png(tfile, width=22, height=20, units='cm', res=600)
SOmap(Trim=-45, Grats=TRUE)
dev.off()
unlink(tfile)
 SOmap(Trim = -45, Grats = TRUE)
}
}
