% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOplot.R
\name{SOplot}
\alias{SOplot}
\title{Southern Ocean plot}
\usage{
SOplot(x, y = NULL, target = NULL, ..., source = NULL, add = TRUE)
}
\arguments{
\item{x}{longitude vector, or object with coordinates}

\item{y}{lattitude vector, or missing if x is an object}

\item{target}{target projection (default = stereo)}

\item{...}{other plot options}

\item{source}{starting projection (default = longlat)}

\item{add}{add layer to plot (default = TRUE)}
}
\value{
Produces at the very base a round bathymetry map of the southern hemisphere.
}
\description{
Reproject and add point layers to either SOmap or SOauto_map layers.
}
\examples{
\dontrun{
 x<-c(-70, -60,-50, -90)
 y<-c(-50, -75, -45, -60)
 map<-SOauto_map(x,y, input_lines = FALSE)
 map
 SOplot(x = x, y = y, target = map$projection,pch=19,col=6)
}
}
