% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOproj.R
\name{SOproj}
\alias{SOproj}
\title{Southern projection}
\usage{
SOproj(x, y = NULL, target = NULL, data, ..., source = NULL)
}
\arguments{
\item{x}{longitude vector, or object with coordinates}

\item{y}{lattitude vector}

\item{target}{target projection (default = stereo)}

\item{data}{optional data to be included}

\item{...}{arguments passed to [reproj::reproj()]}

\item{source}{starting projection (default = longlat)}
}
\value{
Produces at the very base a round bathymetry map of the southern hemisphere.
}
\description{
Function for reprojecting data.
}
\examples{
\dontrun{
 x <- c(-70, -60,-50, -90)
 y <- c(-50, -75, -45, -60)
 pnts <- SOproj(x = y, y = x)
 SOmap2(CCAMLR = TRUE)
 plot(pnts, pch = 19, col = 3, add = TRUE)
}
}
